working on weird memory error with collider filtering

+++ collision.h
@@ -14,7 +14,7 @@ extern uint32_t numColliders;
 
 void initCollision(void);
 Collider* addCollider(uint32_t x1, uint32_t y1, uint32_t x2, uint32_t y2);
-uint16_t getNearbyColliders(uint16_t pos[2], uint8_t radius, Collider** dest);
+uint16_t getNearbyColliders(uint32_t pos[2], uint8_t radius, Collider** dest);
 Collider* moveActor(Actor* a);
 bool testCollision(uint32_t* p1, uint32_t* p2, uint32_t* p3, uint16_t r);
